Small tidy ups
